# Made the get all projects function better