--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Fixed
-
+- Fixed the get all projects query - [#436](https://github.com/DigitalExcellence/dex-backend/issues/436)
 ### Security