@Maxvanhattum 2021-05-10
Hey @TimSnoek123 , I thought @niraymak changed that getall method to those .loads() because he said it improved the performance. Have you and he compared response times of these different ways of retrieving the projects? Maybe you two can talk about this?