@TimSnoek123 2021-05-10
I will test that tomorrow, .load() shouldn't improve the response time from what I know about it, but if it does then I will add that part back. Still the time should still be improved, because that function did some weird things :). @Maxvanhattum